+++ src/platform_impl/linux/x11/ffi.rs
@@ -1,7 +1,7 @@
 use x11_dl::xmd::CARD32;
 pub use x11_dl::{
     error::OpenError, keysym::*, xcursor::*, xinput::*, xinput2::*, xlib::*, xlib_xcb::*,
-    xrandr::*, xrender::*,
+    xpresent::*, xrandr::*, xrender::*,
 };
 
 // Isn't defined by x11_dl